requirements.txt
flet==0.22.*
pyserial==3.5